requirements.txt
mockito==1.0.12
nose==1.3.7
pika==0.11.0
PyYAML==3.12
